Move OperatingSystem.ky from under the java plugin to top level.
Since Operating System class will be used by all plugins, it was
preferable to have it in a top level package accessible by all plugins
instead of just the java plugin.

This is part of a bug where the Jvm and JavaInfo also needs moving to
top level packages.
