+++ src/main/kotlin/com/beust/kobalt/OperatingSystem.kt
@@ -1,4 +1,4 @@
-package com.beust.kobalt.plugin.java
+package com.beust.kobalt
 
 import java.io.File
 import java.util.*
